requirements.txt
amqp==5.0.6
asgiref==3.4.1
astroid==2.8.0
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
django-celery-results==2.2.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
gunicorn==20.1.0
install==1.3.4
isort==5.9.3
kombu==5.1.0
lazy-object-proxy==1.6.0
Markdown==3.2.2
mccabe==0.6.1
pep8==1.7.1
platformdirs==2.3.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
psycopg2-binary==2.9.1
pylint==2.11.1
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
